fix(internal/git/diff.go): retrieve root folder of repository
Function findGitDir(), which looks for the .git directory in the current
directory or any of its parents. It starts from the current working directory
and iteratively checks if there is a .git directory in the current directory or
one of its parent directories. If it finds a .git directory, it returns the
relative path to that directory. If it doesn't find the .git directory after
checking all parent directories, it returns an error.